Added check on occurrence deletion

+++ src/UNL/UCBCN/Manager/DeleteDateTime.php
@@ -5,6 +5,7 @@ use UNL\UCBCN\User;
 use UNL\UCBCN\Calendar;
 use UNL\UCBCN\Event;
 use UNL\UCBCN\Event\Occurrence;
+use UNL\UCBCN\Event\Occurrences as Occurrences;
 use UNL\UCBCN\Permission;
 
 class DeleteDateTime extends PostHandler
@@ -53,9 +54,20 @@ class DeleteDateTime extends PostHandler
             throw new \Exception("The event_datetime_id in the post data must match the event_datetime_id in the URL", 400);
         }
 
+        $event = Event::getById($this->event_datetime->event_id);
+        $allOccurrences = new Occurrences(array(
+            'event_id' => $event->id,
+        ));
+
+        if (count($allOccurrences) === 1) {
+            $this->flashNotice(parent::NOTICE_LEVEL_ALERT, 'Date/Time Delete Error', 'You need at least one Date/Time instance for your event.');
+            // Redirect to the event edit page
+            return $this->event->getEditURL($this->calendar);
+        }
+
         $this->event_datetime->delete();
 
-        $this->flashNotice(parent::NOTICE_LEVEL_SUCCESS, 'Location/Date/Time Deleted', 'A location/date/time for your event ' . $this->event->title . ' has been deleted.');
+        $this->flashNotice(parent::NOTICE_LEVEL_SUCCESS, 'Date/Time Deleted', 'A Date/Time for your event ' . $this->event->title . ' has been deleted.');
 
         // Redirect to the event edit page
         return $this->event->getEditURL($this->calendar);
